LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/text-indent-expected.txt


PASS Can set 'text-indent' to CSS-wide keywords
FAIL Can set 'text-indent' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'text-indent' to a length assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSUnitValue]"
FAIL Can set 'text-indent' to a percent assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSUnitValue]"
PASS Setting 'text-indent' to a time throws TypeError
PASS Setting 'text-indent' to an angle throws TypeError
PASS Setting 'text-indent' to a flexible length throws TypeError
FAIL Setting 'text-indent' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'text-indent' to a URL throws TypeError
PASS Setting 'text-indent' to a transform throws TypeError
FAIL 'text-indent' does not supported '5em each-line' assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
FAIL 'text-indent' does not supported '5em hanging' assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
FAIL 'text-indent' does not supported '5em hanging each-line' assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
